backend/pyproject.toml
[project]
name = "metabodash-backend"
version = "0.1.0"
description = "MetaboDash backend API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "pyjwt>=2.8",
  "cryptography>=42.0",
  "passlib[bcrypt]>=1.7",
  "bcrypt>=3.2,<4.1",
  "httpx[http2]>=0.27",
  "numpy>=1.26",
  "jsonschema>=4.21",
  "redis>=5.0",
  "celery>=5.3",
  "boto3>=1.34",
  "python-dateutil>=2.9",
  "orjson>=3.10",
  "sse-starlette>=2.1",
  "email-validator>=2.1",
  "openai>=1.12",
  "pandas>=2.1",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
